requirements.txt
astroid==2.12.4
click==8.1.3
dill==0.3.5.1
Flask==2.2.2
gunicorn==20.1.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pylint==2.15.0
tomli==2.0.1
tomlkit==0.11.4
Werkzeug==2.2.2
wrapt==1.14.1

PyGithub~=1.55